requirements.txt
pycryptodome
colorama
bs4
brotli